+++ snapshot-tests-documentation/src/test/java/de/skuzzle/test/snapshots/snippets/ChangeNames.java
@@ -2,7 +2,7 @@ package de.skuzzle.test.snapshots.snippets;
 
 import org.junit.jupiter.api.Test;
 
-import de.skuzzle.test.snapshots.SnapshotDsl.Snapshot;
+import de.skuzzle.test.snapshots.Snapshot;
 import de.skuzzle.test.snapshots.SnapshotNaming;
 import de.skuzzle.test.snapshots.junit5.EnableSnapshotTests;