+++ snapshot-tests-documentation/src/test/java/de/skuzzle/test/snapshots/snippets/ParameterizedTests.java
@@ -3,7 +3,7 @@ package de.skuzzle.test.snapshots.snippets;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import de.skuzzle.test.snapshots.SnapshotDsl.Snapshot;
+import de.skuzzle.test.snapshots.Snapshot;
 import de.skuzzle.test.snapshots.SnapshotNaming;
 import de.skuzzle.test.snapshots.junit5.EnableSnapshotTests;