+++ snapshot-tests-documentation/src/test/java/de/skuzzle/test/snapshots/snippets/QuickstartTest.java
@@ -3,8 +3,8 @@ package de.skuzzle.test.snapshots.snippets;
 // tag::quickstart[]
 import org.junit.jupiter.api.Test;
 
-import de.skuzzle.test.snapshots.EnableSnapshotTests;
-import de.skuzzle.test.snapshots.SnapshotDsl.Snapshot;
+import de.skuzzle.test.snapshots.Snapshot;
+import de.skuzzle.test.snapshots.junit5.EnableSnapshotTests;
 
 @EnableSnapshotTests // <1>
 public class QuickstartTest {