+++ snapshot-tests-documentation/src/test/java/de/skuzzle/test/snapshots/snippets/UpdatingSnapshots.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 import de.skuzzle.test.snapshots.ForceUpdateSnapshots;
-import de.skuzzle.test.snapshots.SnapshotDsl.Snapshot;
+import de.skuzzle.test.snapshots.Snapshot;
 import de.skuzzle.test.snapshots.junit5.EnableSnapshotTests;
 
 @Disabled